Update transfer-bar-add.js
Adjusting the final transferBubble to match the format of the other bubbles.  I originally had it switched where I was showing how many chats all operators have, but it should actually be showing the number of available spaces.  This change takes that into account. 

For example, if I have an availability of 3 and one current chat:

Original: 1/3
Change: 2/3